Add Plate options to OMERO side UI  +  align OMERO & FILE menus as far as possible NB Load Acceptor removed from OMERO side until we have test data.

+++ FLIMfitFrontEnd/Classes/@OMERO_data_series/OMERO_data_series.m
@@ -70,12 +70,10 @@ classdef OMERO_data_series < flim_data_series
         
         FLIM_modality;
         
-       
-        datasetId = java.lang.Long(-1);
-        datasetForOutputId = java.lang.Long(-1);
-        plateId = java.lang.Long(-1);
-        
-        
+        datasetId = -1;
+        datasetForOutputId = -1;
+        plateId = -1;
+           
     end
     
     events
